move cart tests to page header
